% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{calc_coverage_chr}
\alias{calc_coverage_chr}
\title{Calculate coverage of a BSseq object for a given chromosome}
\usage{
calc_coverage_chr(chr, bs, min_cov = 1)
}
\arguments{
\item{chr}{single chromosome}

\item{bs}{BSseq object}

\item{min_cov}{minimum value to count base as covered}
}
\value{
data frame with the following columns:
\itemize{
\item \code{sample} sample names
\item \code{chr} chromosome
\item \code{min_cov} minimum coverage threshold (same as \code{min_cov} parameter)
\item \code{covered} number of covered bases
\item \code{total} number of bases
}
}
\description{
This function calculates coverage of bases of a single chromosome
(reference name, \code{seqnames}).
}
\examples{
library(bsseqData)
data(BS.cancer.ex)
calc_coverage_chr("chr21", bs = BS.cancer.ex, min_cov = 10)
}
