% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{calc_coverage_metrics}
\alias{calc_coverage_metrics}
\title{Calculate coverage metrics of a given BSseq object}
\usage{
calc_coverage_metrics(bs, chrs, regions_list, min_cov = 1)
}
\arguments{
\item{bs}{BSseq object}

\item{chrs}{zero or more chromosomes}

\item{regions_list}{named list with zero or more regions}

\item{min_cov}{}
}
\value{
named list of the following data frames:
\itemize{
\item \code{cov} genome-wide coverage metrics (see \code{\link[=calc_coverage]{calc_coverage()}})
\item \code{cov_chr} metrics summarized by chromosomes (see \code{\link[=calc_coverage_chr]{calc_coverage_chr()}})
\item \code{cov_regions} metrics summarized by regions (see \code{\link[=calc_coverage_regions]{calc_coverage_regions()}})
}
}
\description{
This is the main function to calculate coverage metrics.
It inputs a \link[bsseq:BSseq]{bsseq::BSseq} object, zero or more chromosomes
(reference names, \code{seqnames}), and zero or more regions.
}
\details{
The \code{regions_list} must be a names list of data frames and
\link[GenomicRanges:GRanges-class]{GenomicRanges::GRanges} objects. Names of list elements are used to
differentiate metrics from one set of regions to others (column \code{type}).

If \code{chrs} or \code{regions_list} is not provided it will not calculate for these
genomic intervals.
}
\examples{
library(bsseqData)
data(BS.cancer.ex)
calc_coverage_metrics(BS.cancer.ex,
  chrs = c("chr21","chr22"),
  regions_list = list(custom = GRanges(seqnames = "chr22", ranges = IRanges(start = 1, end = 2*10^7))),
  min_cov = 10)
}
