% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{calc_coverage_regions}
\alias{calc_coverage_regions}
\title{Calculate coverage of a BSseq object where bases overlap given regions}
\usage{
calc_coverage_regions(regions, type, bs, min_cov = 1)
}
\arguments{
\item{regions}{data frame or GenomicRanges}

\item{type}{region type (exon, intron, gene symbol, etc)}

\item{bs}{BSseq object}

\item{min_cov}{minimum value to count base as covered}
}
\value{
data frame with the following columns:
\itemize{
\item \code{sample} sample names
\item \code{type} region type
\item \code{min_cov} minimum coverage threshold (same as \code{min_cov} parameter)
\item \code{covered} number of covered bases
\item \code{total} number of bases
}
}
\description{
This function inputs a \link[bsseq:BSseq]{bsseq::BSseq} object and a set of regions
(data frame or \link{GRanges}). Bases that overlaps one of these regions
are considered for coverage calculation.
}
\details{
Regions as data frame must have the columns: \code{seqnames}, \code{start} and \code{end}.
}
\examples{
library(bsseqData)
data(BS.cancer.ex)
calc_coverage_regions(
  regions = GRanges(seqnames = "chr22", ranges = IRanges(start = 1, end = 2*10^7)),
  type = "custom",
  bs = BS.cancer.ex,
  min_cov = 10)
}
