% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bismark.R
\name{plot_bismark_nucleotide_coverage}
\alias{plot_bismark_nucleotide_coverage}
\title{Plot Bismark nucleotide coverage}
\usage{
plot_bismark_nucleotide_coverage(
  nucleotide_coverage,
  show = c("Nucleotide", "Dinucleotide")
)
}
\arguments{
\item{nucleotide_coverage}{}

\item{show}{}
}
\value{

}
\description{
Plot Bismark nucleotide coverage
}
