% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bismark.R
\name{plot_genomic_nucleotide_frequencies}
\alias{plot_genomic_nucleotide_frequencies}
\title{Plot genomic nucleotide frequencies}
\usage{
plot_genomic_nucleotide_frequencies(
  genomic_nucleotide_frequencies,
  show = c("Nucleotide", "Dinucleotide")
)
}
\arguments{
\item{genomic_nucleotide_frequencies}{data frame should have the columns:
\itemize{
\item Bases
\item Type
\item Count
}}
}
\value{

}
\description{
Plot genomic nucleotide frequencies
}
