% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bismark.R
\name{plot_mbias}
\alias{plot_mbias}
\title{Plot M-Bias}
\usage{
plot_mbias(mbias)
}
\arguments{
\item{mbias}{data frame must contains the folowing columns:
\itemize{
\item Context
\item Strand (filled with NA if single-end data)
\item Cycle
\item Percent.Methylation
}}
}
\value{
ggplot2 object
}
\description{
This function inputs a data frame generated by read_bismark_mbias function
and plots M-Bias using ggplot2.
}
\details{
If there are Strand information (Forward and Reverse), this function plots a
Strand as facet.
}
