% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bismark.R
\name{read_bismark_align_report}
\alias{read_bismark_align_report}
\title{Read Bismark alignment report file}
\usage{
read_bismark_align_report(file)
}
\arguments{
\item{file}{Path to Bismark alignment report file}
}
\value{
data frame with the following columns:
\itemize{
\item Reads:
\item Unique.Hits:
\item No.Hits:
\item Multiple.Hits:
\item Not.Extracted:
\item OT:
\item OB:
\item CTOT:
\item CTOB:
\item C:
\item CpG.Methylated:
\item CHG.Methylated:
\item CHH.Methylated:
\item Unknown.Methylated:
\item CpG.Unmethylated:
\item CHG.Unmethylated:
\item CHH.Unmethylated:
\item Unknown_Unmethylated:
}
}
\description{
Read Bismark alignment report file
}
