% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bismark.R
\name{read_bismark_deduplicate_report}
\alias{read_bismark_deduplicate_report}
\title{Read Bismark deduplicate report file}
\usage{
read_bismark_deduplicate_report(file)
}
\arguments{
\item{file}{}
}
\value{

}
\description{
Read Bismark deduplicate report file
}
