% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bismark.R
\name{read_bismark_genomic_nucleotide_frequencies}
\alias{read_bismark_genomic_nucleotide_frequencies}
\title{Read Genomic Nucleotide Frequencies}
\usage{
read_bismark_genomic_nucleotide_frequencies(file)
}
\arguments{
\item{file}{path to genomic_nucleotide_frequencies.txt}
}
\value{
data frame with the following columns:
\itemize{
\item Bases: (di-)nucleotides
\item Count: frequency in genome
}
}
\description{
The 'genomic_nucleotide_frequencies.txt' can be generated using either
\code{bismark_genome_preparation --genomic_composition} or \code{bam2nuc} Bismark
commands.
}
