% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bismark.R
\name{read_bismark_mbias}
\alias{read_bismark_mbias}
\title{Read Bismark M-Bias file}
\usage{
read_bismark_mbias(file)
}
\arguments{
\item{file}{Path to Bismark M-Bias file}
}
\value{
data frame with the following columns:
\itemize{
\item Context: CpG, CHG or CHH
\item Strand: Forward or Reverse
\item Cycle: Sequencing cycle
\item Count.Methylated:
\item Count.Unmethylated:
\item Percent.Methylation:
\item Coverage:
}
}
\description{
This function inputs a M-Bias file generated by bismark_methylation_extractor
and returns a data frame (long format) with all contexts and strands found.
}
