% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bismark.R
\name{read_bismark_methylation_extractor}
\alias{read_bismark_methylation_extractor}
\title{Read Bismark methylation extractor report file}
\usage{
read_bismark_methylation_extractor(file)
}
\arguments{
\item{file}{}
}
\value{

}
\description{
Read Bismark methylation extractor report file
}
