% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bismark.R
\name{read_bismark_nucleotide_coverage}
\alias{read_bismark_nucleotide_coverage}
\title{Read Bismark nucleotide coverage}
\usage{
read_bismark_nucleotide_coverage(file)
}
\arguments{
\item{file}{}
}
\value{

}
\description{
Read Bismark nucleotide coverage
}
