% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylcx.R
\name{read_methycx_mbias}
\alias{read_methycx_mbias}
\title{Read methylcx M-Bias file}
\usage{
read_methycx_mbias(file)
}
\arguments{
\item{file}{path M-Bias file}
}
\value{
data frame with the following columns:
\itemize{
\item Context: CpG, CHG or CHH
\item Strand: Forward or Reverse
\item Cycle: Sequencing cycle
\item Count.Methylated:
\item Count.Unmethylated:
\item Percent.Methylation:
\item Coverage:
}
}
\description{
Read methylcx M-Bias file
}
