% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_galore.R
\name{read_trim_galore_stats}
\alias{read_trim_galore_stats}
\title{Read TrimGalore stats file}
\usage{
read_trim_galore_stats(file)
}
\arguments{
\item{file}{stats file}
}
\value{

}
\description{
Read TrimGalore stats file
}
