% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{write_coverage_metrics}
\alias{write_coverage_metrics}
\title{Write coverage metrics data as tabular files}
\usage{
write_coverage_metrics(
  cov_metrics,
  prefix_path = "",
  delim = ",",
  ext = ".csv"
)
}
\arguments{
\item{cov_metrics}{named list of metrics data (see \code{\link[=calc_coverage_metrics]{calc_coverage_metrics()}})}

\item{prefix_path}{single string to add to base file name as prefix}

\item{delim}{delimiter used to separate values. Default is ",".}

\item{ext}{file extension. Default is ".csv".}
}
\value{
file paths to created files.
Possible files are (prefix and ext omitted):
\itemize{
\item \code{cov} genome-wide coverage metrics (see \code{\link[=calc_coverage]{calc_coverage()}})
\item \code{cov_chr} metrics summarized by chromosomes (see \code{\link[=calc_coverage_chr]{calc_coverage_chr()}})
\item \code{cov_regions} metrics summarized by regions (see \code{\link[=calc_coverage_regions]{calc_coverage_regions()}})
}
}
\description{
This function inputs the result object of \code{\link[=calc_coverage_metrics]{calc_coverage_metrics()}}.
Each data frame in the named list is written as tabular file (with header).
The base name of each file is defined by name of element plus \code{ext}.
Some of files may not create if there are missing metrics results,
for instance no regions were passed to \code{calc_coverage_metrics} function.
}
\details{
The \code{prefix_path} are added before base name of each file.
It can be the cytosine context for example (CpG, CHG, CHH).
It also can be use to change destination directory (example: \verb{metrics/}).
}
\examples{
\dontrun{
library(bsseqData)
data(BS.cancer.ex)
calc_coverage_metrics(BS.cancer.ex,
  chrs = c("chr21","chr22"),
  regions_list = list(custom = GRanges(seqnames = "chr22", ranges = IRanges(start = 1, end = 2*10^7))),
  min_cov = 10) \%>\%
  write_coverage_metrics(prefix = "metrics/CpG_")
}
}
